Altered to create a season file for each team Stores each game of the season in the same csv
Added a csv reader to check the csv contents
Needs to be in the individual team folder
to work

+++ GameResultsCall.py
@@ -1,3 +1,4 @@
+import csv
 
 import requests
 import json
@@ -6,7 +7,7 @@ import os
 
 
 # Authentication credentials for API request
-headers = {"Authorization": "Your token here"}
+headers = {"Authorization": "Your Token Here"}
 
 # Convert to readable format
 def jprint(obj):
@@ -20,13 +21,19 @@ def jprint(obj):
 #seasonType = input("Enter 'regular' or 'postseason': ")
 #team = input("Enter team name: ")
 
-# Parameters to be inserted into request
+# Hardcoded parameters for testing
+year = 2023
+week = 1
+seasonType = 'regular'
+team = 'Team Name Here'
+
 parameters = {
-    "year": 2023,
-    "week": 1,
-    "seasonType": "regular",
-    "team": "alabama"
+    "year": year,
+    "week": week,
+    "seasonType": seasonType,
+    "team": team
 }
+
 
 # Function to call game results from API
 def GameResults():
@@ -36,48 +43,67 @@ def GameResults():
     # Convert json to human readable format
     jprint(response.json())
 
+    # Deprecated; Used for testing
     df = pd.DataFrame(columns=['GameID', 'Home_Team', 'Away_Team'])
 
 
     # Get list
     data = response.json()
-    # Extract data for database
     for d in data:
         home_team = d['home_team']
-        home_score = d['home_points']
-        away_team = d['away_team']
-        away_score = d['away_points']
-        away_conference = d['away_conference']
-        if home_score > away_score:
-            outcome = 'W'
-        else:
-            outcome = 'F'
-        margin = home_score - away_score
-        date = d['start_date'][:10]
-        game_id = d['id']
+    home_team = home_team.lower()
+    # Extract data for database
+    if home_team == team:
+        for d in data:
+            home_team = d['home_team']
+            home_score = d['home_points']
+            away_team = d['away_team']
+            away_score = d['away_points']
+            away_conference = d['away_conference']
+    else:
+        for d in data:
+            home_team = d['away_team']
+            home_score = d['away_points']
+            away_team = d['home_team']
+            away_score = d['home_points']
+            away_conference = d['home_conference']
 
+    if home_score > away_score:
+        outcome = 'W'
+    else:
+        outcome = 'F'
+    margin = home_score - away_score
+    season = str(d['season'])
+    #date = d['start_date'][:10]
+    game_id = d['id']
+
+        # Deprecated; Used to check output is correct
         #rw = {'GameID': game_id, 'Home_Team': home_score, 'Away_Team': away_score}
         #df.loc[len(df)] = rw
-        df = df.append({'GameID': game_id, 'Home_Team': home_score, "Away_Team": away_score}, ignore_index=True)
+        #df = df.append({'GameID': game_id, 'Home_Team': home_score, "Away_Team": away_score}, ignore_index=True)
 
-    df.columns = ['GameID', home_team, away_team]
+    #df.columns = ['GameID', home_team, away_team]
 
     # Check if the team folder exists, create if not
-    folder_name = os.path.join("Teams", home_team)
-    if not os.path.exists(folder_name):
-        os.makedirs(folder_name)
+    folderName = os.path.join("Teams", team)
+    if not os.path.exists(folderName):
+        os.makedirs(folderName)
 
-    filename = os.path.join(folder_name, f"{date}.csv")
-    # If the file already exists, load, add the row, save
+    filename = os.path.join(folderName, season)
+    # If the file already exists:
     if os.path.exists(filename):
-        # Create a new row from the current game data
-        newFrame = {"Opponent": away_team, "Conference": away_conference, "Outcome": outcome, "Margin": margin}
+        # Read the file and create the new row
         existing = pd.read_csv(filename)
-        update = existing.append(newFrame, ignore_index=True)
-        update.to_csv(filename, index=False)
-    # If not, create, add the row, save
+        newFrame = pd.DataFrame(
+            {"Opponent": [away_team], "Conference": [away_conference], "Outcome": [outcome], "Margin": [margin]})
+        # Check if the new row is already in the file; If not, add it and save
+        row_exists = newFrame.isin(existing).all(axis=1).any()
+        if not row_exists:
+            update = existing.append(newFrame, ignore_index=True)
+            update.to_csv(filename, index=False)
+    # If the file doesn't exist:
     else:
-        # Create a new row from the current game data
+        # Create a new row from the current game data and save as a csv
         newFrame = pd.DataFrame(
             {"Opponent": [away_team], "Conference": [away_conference], "Outcome": [outcome], "Margin": [margin]})
         newFrame.to_csv(filename, index=False)
